+++ src/main/java/com/hotels/bdp/circustrain/bigquery/extraction/BigQueryExtractionData.java
@@ -15,11 +15,7 @@
  */
 package com.hotels.bdp.circustrain.bigquery.extraction;
 
-import static com.hotels.bdp.circustrain.bigquery.extraction.BigQueryDataExtractionKey.makeKey;
-
 import java.util.UUID;
-
-import com.google.cloud.bigquery.Table;
 
 class BigQueryExtractionData {
 
@@ -27,19 +23,6 @@ class BigQueryExtractionData {
   private final String dataBucket = "circus-train-bigquery-tmp-" + UUID.randomUUID().toString().toLowerCase();
   private final String dataKey = UUID.randomUUID().toString().toLowerCase();
   private final String dataUri = "gs://" + dataBucket + "/" + dataKey + "." + format;
-  private final Table table;
-
-  BigQueryExtractionData(Table table) {
-    this.table = table;
-  }
-
-  String getTableName() {
-    return table.getTableId().getTable();
-  }
-
-  String getDatasetName() {
-    return table.getTableId().getDataset();
-  }
 
   String getDataKey() {
     return dataKey;
@@ -55,42 +38,5 @@ class BigQueryExtractionData {
 
   String getDataBucket() {
     return dataBucket;
-  }
-
-  Table getTable() {
-    return table;
-  }
-
-  @Override
-  public int hashCode() {
-    return (makeKey(getDatasetName(), getTableName())).hashCode();
-  }
-
-  @Override
-  public boolean equals(Object obj) {
-    if (this == obj) {
-      return true;
-    }
-    if (obj == null) {
-      return false;
-    }
-    if (getClass() != obj.getClass()) {
-      return false;
-    }
-    BigQueryExtractionData other = (BigQueryExtractionData) obj;
-
-    String thisDatasetName = this.getDatasetName();
-    String otherDatasetName = other.getDatasetName();
-    if (!thisDatasetName.equals(otherDatasetName)) {
-      return false;
-    }
-
-    String thisTableName = this.getTableName();
-    String otherTableName = other.getTableName();
-    if (!thisTableName.equals(otherTableName)) {
-      return false;
-    }
-
-    return true;
   }
 }