+++ src/test/java/com/hotels/bdp/circustrain/bigquery/listener/BigQueryReplicationListenerTest.java
@@ -15,9 +15,11 @@
  */
 package com.hotels.bdp.circustrain.bigquery.listener;
 
+import static org.mockito.Matchers.any;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.verify;
+import static org.mockito.Mockito.verifyZeroInteractions;
 import static org.mockito.Mockito.when;
 
 import org.junit.Before;
@@ -60,17 +62,11 @@ public class BigQueryReplicationListenerTest {
     when(bigQuery.getDataset(anyString())).thenReturn(dataset);
     Table table = mock(Table.class);
     when(dataset.get(anyString())).thenReturn(table);
-    listener.tableReplicationStart(eventTableReplication, "eventId");
-    verify(dataExtractionManager).extract(table);
-  }
 
-  @Test(expected = CircusTrainException.class)
-  public void tableReplicationStartTableMissingThrowsException() {
-    Dataset dataset = mock(Dataset.class);
-    when(bigQuery.getDataset(anyString())).thenReturn(dataset);
-    when(dataset.get(anyString())).thenReturn(null);
     listener.tableReplicationStart(eventTableReplication, "eventId");
 
+    verify(dataExtractionManager).register(any(Table.class));
+    verifyZeroInteractions(storage);
   }
 
   @Test
@@ -81,7 +77,7 @@ public class BigQueryReplicationListenerTest {
     Table table = mock(Table.class);
     when(dataset.get(anyString())).thenReturn(table);
     listener.tableReplicationSuccess(eventTableReplication, "eventId");
-    verify(dataExtractionManager).cleanup(table);
+    verify(dataExtractionManager).cleanup();
   }
 
   @Test
@@ -92,6 +88,6 @@ public class BigQueryReplicationListenerTest {
     Table table = mock(Table.class);
     when(dataset.get(anyString())).thenReturn(table);
     listener.tableReplicationFailure(eventTableReplication, "eventId", mock(CircusTrainException.class));
-    verify(dataExtractionManager).cleanup(table);
+    verify(dataExtractionManager).cleanup();
   }
 }