+++ src/test/java/com/hotels/bdp/circustrain/bigquery/metastore/BigQueryMetastoreClientTest.java
@@ -113,9 +113,8 @@ public class BigQueryMetastoreClientTest {
     String dbName = "database";
     String tableName = "table";
     TableId tableId = TableId.of(dbName, tableName);
-    BigQueryDataExtractionService service = mock(BigQueryDataExtractionService.class);
     String location = "gs://foo/baz";
-    when(dataExtractionManager.location(any(Table.class))).thenReturn(location);
+    when(dataExtractionManager.location()).thenReturn(location);
     when(bigQuery.getDataset(anyString())).thenReturn(dataset);
     when(dataset.get(anyString())).thenReturn(table);
     TableDefinition tableDefinition = mock(TableDefinition.class);