+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # TBD
 ### Changed
 * Improved error logging when extraction job fails or is no longer present.
+* Extract table during replication rather than in listener to prevent silent failure. See [#5](https://github.com/HotelsDotCom/circus-train-bigquery/issues/5).
 
 # 1.0.1 - 2018-02-28
 ### Changed